requirements.txt
# Web framework
fastapi
uvicorn
python-multipart

# Image processing
opencv-python
pillow
numpy

# MediaPipe (face detection + landmarks)
mediapipe

# ML models
torch
torchvision
timm
ultralytics
transformers

# HTTP client (frontend → backend)
requests
